feat: add SkillRadarChart component and integrate it into StatsCharts and ChartTabs
