+++ src/modules/stats/components/chart/index.ts
@@ -3,6 +3,7 @@ export type { ChartData, UserStats } from "../../types/stats.types";
 
 export { SkillExperienceChart } from "./skill-experience-chart/SkillExperienceChart";
 export { LevelProgressionChart } from "./level-progression-chart/LevelProgressionChart";
+export { SkillRadarChart } from "./skill-radar-chart/SkillRadarChart";
 
 // Mock Data
 export * from "../../data/mock-data";
