+++ src/modules/stats/components/chart/level-progression-chart/LevelProgressionChart.tsx
@@ -13,20 +13,22 @@ export const LevelProgressionChart = () => {
   const chartOptions = getChartOptions(levelProgressionData);
 
   return (
-    <div className="w-full rounded-xl bg-[#131928] p-6 shadow-lg">
-      <div className="mb-4">
-        <h3 className="text-lg font-semibold text-white mb-2">Progression des Niveaux</h3>
-        <p className="text-sm text-gray-400">Évolution de votre progression XP et seuils de niveaux</p>
+    <div className="w-full h-full rounded-xl bg-[#131928] p-4 lg:p-6 shadow-lg flex flex-col">
+      <div className="flex-shrink-0 mb-4">
+        <h3 className="text-base lg:text-lg font-semibold text-white mb-2">Progression des Niveaux</h3>
+        <p className="text-xs lg:text-sm text-gray-400">Évolution de votre progression XP et seuils de niveaux</p>
       </div>
 
-      <ReactECharts
-        option={chartOptions}
-        style={{ height: "400px", width: "100%" }}
-        opts={{
-          renderer: "svg",
-          devicePixelRatio: window.devicePixelRatio || 2,
-        }}
-      />
+      <div className="flex-1 min-h-0">
+        <ReactECharts
+          option={chartOptions}
+          style={{ height: "100%", width: "100%" }}
+          opts={{
+            renderer: "svg",
+            devicePixelRatio: window.devicePixelRatio || 2,
+          }}
+        />
+      </div>
     </div>
   );
 };
