+++ src/shared/types/skill.type.ts
@@ -14,6 +14,7 @@ export type Skill = {
   progressValue?: number;
   imageUrl?: string;
   totalXp?: number;
+  totalQuests?: number;
   completedQuests?: number;
   averageQuestXp?: number;
   color?: string;
